requirements.txt
djangorestframework==3.12.1
django-environ==0.4.5
boto3==1.16.7
botocore==1.19.7
